Update Statistics for TeacherUI
